requirements.txt
Django==2.2.10
dj-database-url==0.5.0
django-stubs==1.4.0
python-dotenv==0.11.0
psycopg2_binary==2.8.4
djangorestframework==3.11.0
drf-flex-fields==0.7.5
django-filter==2.2.0
